# docs/swagger.yaml
basePath: /
definitions:
  api.BackfillRequest:
    properties:
      from:
        example: "2026-07-01T00:00:00Z"
        type: string
      to:
        example: "2026-08-01T00:00:00Z"
        type: string
    type: object
  api.CreateDeviceRequest:
    properties:
      device_type_id:
        type: string
      name:
        type: string
    type: object
...
  devices.Device:
    properties:
      device_type_id:
        type: string
      id:
        type: string
      local_id:
        type: string
      name:
        type: string
    type: object
  devices.DeviceType:
    properties:
      id:
        type: string
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/devices.Service'
        type: array
    type: object
  devices.Service:
    properties:
      characteristic_id:
        description: |-
          CharacteristicId and ValuePath come from the output's content variable:
          the characteristic gives the value its meaning and unit, the path names
          the variable inside the message.
        type: string
      direction:
        $ref: '#/definitions/domain.Direction'
      id:
        type: string
      name:
        type: string
      value_path:
        type: string
    type: object
...
  domain.Direction:
    enum:
    - sensor
    - actuator
    type: string
    x-enum-varnames:
    - Sensor
    - Actuator
...
  runtime.BackfillChannelStatus:
    properties:
      asset_id:
        type: string
      backfillable:
        description: Backfillable is false when SkipReason says why not.
        type: boolean
      channel_id:
        type: string
      failed:
        type: integer
      last_error:
        description: |-
          LastError is the most recent publish failure of this channel, kept so a
          job that mostly worked still says what went wrong.
        type: string
      name:
        type: string
      published:
        description: |-
          Published counts the readings that reached the platform, Silent the steps
          that sent nothing - a tick that produced no value at all (a dataset
          outside its own time range), or one whose value did not move far enough
          for a channel publishing on change - and Failed the ones the platform
          refused. The three add up to the steps of the channel's grid.
        type: integer
      silent:
        type: integer
      skip_reason:
        type: string
    type: object
  runtime.BackfillState:
    enum:
    - running
    - done
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - BackfillRunning
    - BackfillDone
    - BackfillFailed
    - BackfillCancelled
  runtime.BackfillStatus:
    properties:
      channels:
        items:
          $ref: '#/definitions/runtime.BackfillChannelStatus'
        type: array
      channels_done:
        type: integer
      channels_total:
        description: |-
          ChannelsTotal counts every channel of the environment, backfillable or
          not; ChannelsDone counts the ones that are finished with.
        type: integer
      current_channel:
        description: CurrentChannel and Position are where the job stands right now.
        type: string
      environment_id:
        type: string
      error:
        type: string
      finished_at:
        type: string
      from:
        type: string
      position:
        type: string
      published:
        type: integer
      started_at:
        type: string
      state:
        $ref: '#/definitions/runtime.BackfillState'
      to:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Simulates environments — sites, buildings and apartments — and
    their devices, and publishes their data as if it came from real
    hardware. An environment is one document: GET returns what PUT
    accepts, so a whole site can be created in a single call.

    The legacy world/room/device endpoints are being replaced by the
    environment api and are not documented here.
  title: MOSES API
  version: 1.0.0
paths:
...
  /device-types:
    get:
      description: 'Every device type that publishes through this service''s protocol,
        with its services: the id, name, direction and characteristic of each, which
        is everything the channels of an asset need. Offering this list is what lets
        an editor ask for a device type instead of an id.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/devices.DeviceType'
            type: array
        "401":
          description: the token carries no subject
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: List the device types an asset can be built from
      tags:
      - Catalog
  /devices:
    post:
      consumes:
      - application/json
      description: Registers a device of the given type and returns its id, which
        becomes the asset's external_ref. It is created with the caller's own token,
        so it belongs to whoever built the asset.
      parameters:
      - description: device type and name
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/api.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/devices.Device'
        "400":
          description: the body is unreadable or incomplete
          schema:
            type: string
        "401":
          description: the token carries no subject
          schema:
            type: string
        "413":
          description: the request body is larger than the allowed limit
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create the platform device of a simulated asset
      tags:
      - Catalog
  /devices/{id}:
    delete:
      description: Removing an asset should not leave its device behind. Deleting
        one that does not exist is not an error.
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: deleted
          schema:
            type: string
        "401":
          description: the token carries no subject
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete the platform device of a simulated asset
      tags:
      - Catalog
...
  /environments/{id}/backfill:
    get:
      description: |-
        Where the job stands: running, done, failed or cancelled, how many channels are finished with, which instant it is at, and per channel either the number of readings published or the reason it was skipped.

        The registry is held in memory only. A restart forgets every job, and this then answers 404 rather than claiming a state it cannot know: a job is not resumable, because it would have to know which readings already reached the platform, and the platform keeps a second one rather than replacing it.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/runtime.BackfillStatus'
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, no access to it, or nothing is known about
            a backfill of it
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: The backfill of one environment
      tags:
      - Environment
    post:
      consumes:
      - application/json
      description: |-
        Computes the environment over the given window and publishes every reading with the timestamp it would have had, so that a model can be trained on history the environment never actually lived through. Runs asynchronously; the response is the job as it stands at that moment, and GET on the same path follows it.

        Only a channel whose platform service declares the attribute `senergy/time_path` can be backfilled, because that attribute is what makes the platform read the event time out of the payload instead of stamping the arrival time. moses does not change device types to add it. Only profile and dataset sources are reconstructed: a script source depends on the state its earlier runs left behind, and a formula follows from other channels. Every channel that is skipped says why in the status.

        The window may not end in the future, may not span more than 366 days, and may not come to more than two million readings in total across the channels of the environment; a reading is published synchronously, so that count is the runtime of the job.
      parameters:
      - description: environment id
        in: path
        name: id
        required: true
        type: string
      - description: the window to reconstruct
        in: body
        name: window
        required: true
        schema:
          $ref: '#/definitions/api.BackfillRequest'
      produces:
      - application/json
      responses:
        "202":
          description: accepted, with the job as it stands
          schema:
            $ref: '#/definitions/runtime.BackfillStatus'
        "400":
          description: the body is unreadable, or the window is empty, in the future,
            too long or too dense
          schema:
            type: string
        "401":
          description: the token carries no subject
          schema:
            type: string
        "404":
          description: no such environment, no access to it, or it is not running
            here
          schema:
            type: string
        "409":
          description: a backfill or a history run of this environment is already
            running
          schema:
            type: string
        "413":
          description: the request body is larger than the allowed limit
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      security:
      - Bearer: []
      summary: Reconstruct one environment over a past window
      tags:
      - Environment
...
securityDefinitions:
  Bearer:
    description: A keycloak issued JWT. Verified at the gateway, not here.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"